Modularized DataRepository and the db implementation
