+++ lib/services/data/db/db_repository.dart
@@ -1,0 +1,8 @@
+import 'package:meta/meta.dart';
+
+import 'mongodb_provider.dart';
+
+abstract class DbRepository {
+	@protected
+	MongoDbProvider get client;
+}
